+++ doc/changelog.md
@@ -3,6 +3,7 @@
 
 ## Work In Progress
 
+- Renamed warnings to notes.
 - Made `wuffs_base__status` a struct.